:label: (business) FindClinicalTrial repository definition
FindClinicalTrials is a port definition of a function that retrieves ClinicalTrial according many properties:
  - the sponsor name
  - the country code
  - the date before the trial start
  - the date after the trial end
As all ports, this is just a definition, not an implementation.
Specific implementations could be provided according to the project need.
For instance it could be a third-party API listing all clinical trials.
But it also could be a query in a database or just a file load.