:label: (business) GetOnGoingClinicalTrials use case definition
GetOnGoingClinicalTrials is a use case definition of a function that retrieves ClinicalTrial according 2 properties:
  - the sponsor name
  - the country code

This function returns only _ongoing_ trials that is:
  - its start date is in the past
  - its end date is in the future
  - it has not been canceled

By definition, the business core of the application should not directly depend on a specific service that returns clinical trials.
That is why an adapter of the FindClinicalTrials is injected to solve this use case.
This adapter will be choose at the application start.